+++ backend/src/main/java/com/nuriwoolim/pagebackend/user/entity/UserType.java
@@ -1,4 +1,4 @@
-package com.nuriwoolim.pagebackend.user;
+package com.nuriwoolim.pagebackend.user.entity;
 
 public enum UserType {
     ADMIN, MANAGER, MEMBER, NONMEMBER
