+++ backend/src/main/java/com/nuriwoolim/pagebackend/user/repository/UserRepository.java
@@ -1,9 +1,10 @@
-package com.nuriwoolim.pagebackend.user;
+package com.nuriwoolim.pagebackend.user.repository;
 
+import com.nuriwoolim.pagebackend.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
-    
+
 }
