+++ backend/src/main/java/com/nuriwoolim/pagebackend/user/service/UserService.java
@@ -1,9 +1,11 @@
-package com.nuriwoolim.pagebackend.user;
+package com.nuriwoolim.pagebackend.user.service;
 
 import com.nuriwoolim.pagebackend.global.exception.CustomException;
 import com.nuriwoolim.pagebackend.global.exception.ErrorCode;
 import com.nuriwoolim.pagebackend.user.dto.UserCreateRequest;
 import com.nuriwoolim.pagebackend.user.dto.UserUpdateRequest;
+import com.nuriwoolim.pagebackend.user.entity.User;
+import com.nuriwoolim.pagebackend.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
