+++ backend/src/test/java/com/nuriwoolim/pagebackend/user/UserServiceTest.java
@@ -10,6 +10,10 @@ import com.nuriwoolim.pagebackend.global.exception.CustomException;
 import com.nuriwoolim.pagebackend.global.exception.ErrorCode;
 import com.nuriwoolim.pagebackend.user.dto.UserCreateRequest;
 import com.nuriwoolim.pagebackend.user.dto.UserUpdateRequest;
+import com.nuriwoolim.pagebackend.user.entity.User;
+import com.nuriwoolim.pagebackend.user.entity.UserType;
+import com.nuriwoolim.pagebackend.user.repository.UserRepository;
+import com.nuriwoolim.pagebackend.user.service.UserService;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
